Add fixed vs floating swap; inherit VanillaSwap
