+++ ql/instruments/vanillaswap.hpp
@@ -26,7 +26,7 @@
 #ifndef quantlib_vanilla_swap_hpp
 #define quantlib_vanilla_swap_hpp
 
-#include <ql/instruments/swap.hpp>
+#include <ql/instruments/fixedvsfloatingswap.hpp>
 #include <ql/time/daycounter.hpp>
 #include <ql/time/schedule.hpp>
 #include <ql/optional.hpp>
@@ -35,7 +35,7 @@ namespace QuantLib {
 
     class IborIndex;
 
-    //! Plain-vanilla swap: fix vs floating leg
+    //! Plain-vanilla swap: fix vs ibor leg
     /*! \ingroup instruments
 
         If no payment convention is passed, the convention of the
@@ -62,11 +62,8 @@ namespace QuantLib {
         - the correctness of the returned value is tested by checking
           it against a known good value.
     */
-    class VanillaSwap : public Swap {
+    class VanillaSwap : public FixedVsFloatingSwap {
       public:
-        class arguments;
-        class results;
-        class engine;
         VanillaSwap(Type type,
                     Real nominal,
                     Schedule fixedSchedule,
@@ -78,139 +75,10 @@ namespace QuantLib {
                     DayCounter floatingDayCount,
                     ext::optional<BusinessDayConvention> paymentConvention = ext::nullopt,
                     ext::optional<bool> useIndexedCoupons = ext::nullopt);
-        //! \name Inspectors
-        //@{
-        Type type() const;
-        Real nominal() const;
-
-        const Schedule& fixedSchedule() const;
-        Rate fixedRate() const;
-        const DayCounter& fixedDayCount() const;
-
-        const Schedule& floatingSchedule() const;
-        const ext::shared_ptr<IborIndex>& iborIndex() const;
-        Spread spread() const;
-        const DayCounter& floatingDayCount() const;
-
-        BusinessDayConvention paymentConvention() const;
-
-        const Leg& fixedLeg() const;
-        const Leg& floatingLeg() const;
-        //@}
-
-        //! \name Results
-        //@{
-        Real fixedLegBPS() const;
-        Real fixedLegNPV() const;
-        Rate fairRate() const;
-
-        Real floatingLegBPS() const;
-        Real floatingLegNPV() const;
-        Spread fairSpread() const;
-        //@}
-        // other
-        void setupArguments(PricingEngine::arguments* args) const override;
-        void fetchResults(const PricingEngine::results*) const override;
 
       private:
-        void setupExpired() const override;
-        Type type_;
-        Real nominal_;
-        Schedule fixedSchedule_;
-        Rate fixedRate_;
-        DayCounter fixedDayCount_;
-        Schedule floatingSchedule_;
-        ext::shared_ptr<IborIndex> iborIndex_;
-        Spread spread_;
-        DayCounter floatingDayCount_;
-        BusinessDayConvention paymentConvention_;
-        // results
-        mutable Rate fairRate_;
-        mutable Spread fairSpread_;
+        void setupFloatingArguments(arguments* args) const override;
     };
-
-
-    //! %Arguments for simple swap calculation
-    class VanillaSwap::arguments : public Swap::arguments {
-      public:
-        arguments() : nominal(Null<Real>()) {}
-        Type type = Receiver;
-        Real nominal;
-
-        std::vector<Date> fixedResetDates;
-        std::vector<Date> fixedPayDates;
-        std::vector<Time> floatingAccrualTimes;
-        std::vector<Date> floatingResetDates;
-        std::vector<Date> floatingFixingDates;
-        std::vector<Date> floatingPayDates;
-
-        std::vector<Real> fixedCoupons;
-        std::vector<Spread> floatingSpreads;
-        std::vector<Real> floatingCoupons;
-        void validate() const override;
-    };
-
-    //! %Results from simple swap calculation
-    class VanillaSwap::results : public Swap::results {
-      public:
-        Rate fairRate;
-        Spread fairSpread;
-        void reset() override;
-    };
-
-    class VanillaSwap::engine : public GenericEngine<VanillaSwap::arguments,
-                                                     VanillaSwap::results> {};
-
-
-    // inline definitions
-
-    inline Swap::Type VanillaSwap::type() const {
-        return type_;
-    }
-
-    inline Real VanillaSwap::nominal() const {
-        return nominal_;
-    }
-
-    inline const Schedule& VanillaSwap::fixedSchedule() const {
-        return fixedSchedule_;
-    }
-
-    inline Rate VanillaSwap::fixedRate() const {
-        return fixedRate_;
-    }
-
-    inline const DayCounter& VanillaSwap::fixedDayCount() const {
-        return fixedDayCount_;
-    }
-
-    inline const Schedule& VanillaSwap::floatingSchedule() const {
-        return floatingSchedule_;
-    }
-
-    inline const ext::shared_ptr<IborIndex>& VanillaSwap::iborIndex() const {
-        return iborIndex_;
-    }
-
-    inline Spread VanillaSwap::spread() const {
-        return spread_;
-    }
-
-    inline const DayCounter& VanillaSwap::floatingDayCount() const {
-        return floatingDayCount_;
-    }
-
-    inline BusinessDayConvention VanillaSwap::paymentConvention() const {
-        return paymentConvention_;
-    }
-
-    inline const Leg& VanillaSwap::fixedLeg() const {
-        return legs_[0];
-    }
-
-    inline const Leg& VanillaSwap::floatingLeg() const {
-        return legs_[1];
-    }
 
 }
